requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
azure-cognitiveservices-speech==1.38.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
multidict==6.0.5
openai==0.28.0
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
